Refactor model picker
