+++ src/vs/sessions/MOBILE.md
@@ -152,7 +152,8 @@ Mobile picker subclasses live in `contrib/` alongside their base classes (not in
 | `contrib/chat/browser/webWorkspacePicker.ts` | `WorkspacePicker` | Web variant: scopes to active host filter and renders as a bottom sheet on phone. Note: this is the only "mobile" picker that lives in a non-`mobile*`-named file because the same class also handles the desktop-web case (host scoping). |
 | `contrib/chat/browser/mobile/mobileWorkspacePickerSheet.ts` | (helper) | Builds `IMobilePickerSheetItem[]` from workspace picker items + browse actions. Used by `WebWorkspacePicker` on phone. |
 | `contrib/providers/agentHost/browser/agentHostSessionConfigPicker.ts` | `AgentHostSessionConfigPicker` | The phone variant `MobileAgentHostSessionConfigPicker` is a private subclass defined **in the same file** as the base (to avoid a circular ESM import); it routes Isolation + Branch to a unified bottom sheet on phone. |
-| `contrib/providers/agentHost/browser/mobile/mobileChatInputConfigPicker.ts` | (standalone) | Phone-only chat-input chip that combines Mode + Model into a unified bottom sheet. Replaces the desktop mode + model pickers (gated off via `when:` clauses) on phone-layout viewports. |
+| `contrib/providers/agentHost/browser/mobile/mobileChatInputConfigPicker.ts` | (standalone) | Phone-only chat-input chip that combines Mode + Model into a unified bottom sheet. It consumes the input-scoped `SessionModelSelectionModel`, so it shares the desktop picker's provider catalog, current selection, canonical persistence, and empty-model state without enumerating language models itself. |
+| `contrib/providers/agentHost/browser/mobile/mobileChatPhoneInputPresenter.ts` | `IChatPhonePresenterImpl` | Builds the combined sheet for an opened chat. Agent Host rows come from `ISessionsProvider.getModels`; every action re-resolves and matches the provider, session, and active-chat identity captured when the sheet opened, then validates the selected id against a fresh provider catalog before writing. |
 
 ### Layout & Navigation
