+++ src/vs/sessions/contrib/providers/agentHost/browser/agentHostModeIcon.ts
@@ -1,0 +1,16 @@
+/*---------------------------------------------------------------------------------------------
+ *  Copyright (c) Microsoft Corporation. All rights reserved.
+ *  Licensed under the MIT License. See License.txt in the project root for license information.
+ *--------------------------------------------------------------------------------------------*/
+
+import { Codicon } from '../../../../../base/common/codicons.js';
+import { ThemeIcon } from '../../../../../base/common/themables.js';
+
+export function getAgentHostModeIcon(value: string | undefined): ThemeIcon | undefined {
+	switch (value) {
+		case 'plan': return Codicon.checklist;
+		case 'autopilot': return Codicon.rocket;
+		case 'interactive': return Codicon.comment;
+		default: return undefined;
+	}
+}
