+++ src/vs/sessions/contrib/providers/agentHost/browser/mobile/mobileChatPhoneInputPresenter.ts
@@ -3,7 +3,6 @@
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
-import { Codicon } from '../../../../../../base/common/codicons.js';
 import { ThemeIcon } from '../../../../../../base/common/themables.js';
 import { Disposable, IDisposable, MutableDisposable } from '../../../../../../base/common/lifecycle.js';
 import { derived, IObservable } from '../../../../../../base/common/observable.js';
@@ -12,34 +11,26 @@ import { SessionConfigKey } from '../../../../../../platform/agentHost/common/se
 import { ICommandService } from '../../../../../../platform/commands/common/commands.js';
 import { IContextKeyService } from '../../../../../../platform/contextkey/common/contextkey.js';
 import { observableContextKey } from '../../../../../../platform/observable/common/platformObservableUtils.js';
-import { IStorageService, StorageScope } from '../../../../../../platform/storage/common/storage.js';
+import { IStorageService } from '../../../../../../platform/storage/common/storage.js';
 import { IInstantiationService } from '../../../../../../platform/instantiation/common/instantiation.js';
 import { IWorkbenchContribution, registerWorkbenchContribution2, WorkbenchPhase } from '../../../../../../workbench/common/contributions.js';
 import { IToggleChatModeArgs, ToggleAgentModeActionId } from '../../../../../../workbench/contrib/chat/browser/actions/chatExecuteActions.js';
 import { IChatPhoneInputPresenter, IChatPhonePresenterImpl } from '../../../../../../workbench/contrib/chat/browser/widget/input/chatPhoneInputPresenter.js';
 import { IModePickerDelegate } from '../../../../../../workbench/contrib/chat/browser/widget/input/modePickerActionItem.js';
-import { IModelPickerDelegate } from '../../../../../../workbench/contrib/chat/browser/widget/input/modelPickerActionItem.js';
-import { getModelProviderIcon } from '../../../../../../workbench/contrib/chat/browser/widget/input/modelProviderIcons.js';
+import { IModelPickerDelegate } from '../../../../../../workbench/contrib/chat/browser/widget/input/modelPicker/modelPickerActionItem.js';
+import { getModelProviderIcon } from '../../../../../../workbench/contrib/chat/browser/widget/input/modelPicker/modelProviderIcons.js';
 import { IChatMode } from '../../../../../../workbench/contrib/chat/common/chatModes.js';
-import { ILanguageModelChatMetadataAndIdentifier, ILanguageModelsService } from '../../../../../../workbench/contrib/chat/common/languageModels.js';
+import { ILanguageModelChatMetadataAndIdentifier } from '../../../../../../workbench/contrib/chat/common/languageModels.js';
 import { IWorkbenchLayoutService } from '../../../../../../workbench/services/layout/browser/layoutService.js';
-import { IChatWidgetService } from '../../../../../../workbench/contrib/chat/browser/chat.js';
-import { isAgentHostProvider } from '../../../../../common/agentHostSessionsProvider.js';
+import { type IAgentHostSessionsProvider, isAgentHostProvider } from '../../../../../common/agentHostSessionsProvider.js';
 import { ISessionsService } from '../../../../../services/sessions/browser/sessionsService.js';
-import { SessionStatus } from '../../../../../services/sessions/common/session.js';
 import { ISessionsProvidersService } from '../../../../../services/sessions/browser/sessionsProvidersService.js';
+import { IActiveSession } from '../../../../../services/sessions/common/sessionsManagement.js';
 import { showMobilePickerSheet, IMobilePickerSheetItem } from '../../../../../browser/parts/mobile/mobilePickerSheet.js';
-import { agentHostModelPickerStorageKey, setAgentHostModelSelection } from '../agentHostModelPicker.js';
-import { isWellKnownModeSchema } from '../agentHostPermissionPickerDelegate.js';
-
-function getAgentHostModeIcon(value: string | undefined): ThemeIcon | undefined {
-	switch (value) {
-		case 'plan': return Codicon.checklist;
-		case 'autopilot': return Codicon.rocket;
-		case 'interactive': return Codicon.comment;
-		default: return undefined;
-	}
-}
+import { getAgentHostModeIcon } from '../agentHostModeIcon.js';
+import { isWellKnownModeSchema, isWellKnownModeValue } from '../agentHostPermissionPickerDelegate.js';
+import { normalizeModelPickerOptions, selectAvailableSessionModel } from '../../../../chat/browser/modelPickerSelection.js';
+import { createChatPhoneInputTarget, IChatPhoneInputTarget, matchesChatPhoneInputTarget } from './mobileChatPhoneInputTarget.js';
 
 /**
  * Action id passed to the workbench `ToggleAgentModeActionId` command when
@@ -51,6 +42,8 @@ type ChatPhonePickerAction =
 	| { kind: 'model'; model: ILanguageModelChatMetadataAndIdentifier }
 	| { kind: 'agentHostMode'; value: string }
 	| { kind: 'agentHostModel'; model: ILanguageModelChatMetadataAndIdentifier };
+
+type RegisterChatPhonePickerAction = (action: ChatPhonePickerAction) => string;
 
 /**
  * Sessions-side implementation of {@link IChatPhoneInputPresenter}.
@@ -73,9 +66,7 @@ class MobileChatPhoneInputPresenter extends Disposable implements IChatPhonePres
 		@IWorkbenchLayoutService private readonly _layoutService: IWorkbenchLayoutService,
 		@ISessionsService private readonly _sessionsService: ISessionsService,
 		@ISessionsProvidersService private readonly _sessionsProvidersService: ISessionsProvidersService,
-		@ILanguageModelsService private readonly _languageModelsService: ILanguageModelsService,
 		@IStorageService private readonly _storageService: IStorageService,
-		@IChatWidgetService private readonly _chatWidgetService: IChatWidgetService,
 	) {
 		super();
 
@@ -105,187 +96,23 @@ class MobileChatPhoneInputPresenter extends Disposable implements IChatPhonePres
 			return id;
 		};
 
-		const sheetItems: IMobilePickerSheetItem[] = [];
-
-		// If the active session is an agent-host session, show its
-		// agent-host config (interactive / plan / autopilot via
-		// {@link SessionConfigKey.Mode}) and the resource-scheme-filtered
-		// model list — the same data path as the empty new-chat input
-		// picker (see `MobileChatInputConfigPicker`). The workbench
-		// chat-modes (Ask / Edit / Agent) and unfiltered model list
-		// surfaced through `modeDelegate`/`modelDelegate` belong to the
-		// default Copilot chat protocol and are wrong for agent-host
-		// sessions.
 		const activeSession = this._sessionsService.activeSession.get();
+		const target = createChatPhoneInputTarget(activeSession);
 		const rawProvider = activeSession ? this._sessionsProvidersService.getProvider(activeSession.providerId) : undefined;
 		const agentHostProvider = rawProvider && isAgentHostProvider(rawProvider) ? rawProvider : undefined;
-
+		let sheetItems: IMobilePickerSheetItem[];
 		if (activeSession && agentHostProvider) {
-			const config = agentHostProvider.getSessionConfig(activeSession.sessionId);
-			const modeSchema = config?.schema.properties[SessionConfigKey.Mode];
-			const modeItems = (modeSchema && isWellKnownModeSchema(modeSchema))
-				? (modeSchema.enum ?? []).map((value, index) => ({
-					value: String(value),
-					label: modeSchema.enumLabels?.[index] ?? String(value),
-					description: modeSchema.enumDescriptions?.[index],
-				}))
-				: [];
-			const rawCurrentMode = config?.values[SessionConfigKey.Mode] ?? modeSchema?.default;
-			const currentModeValue = (typeof rawCurrentMode === 'string' && modeItems.some(i => i.value === rawCurrentMode))
-				? rawCurrentMode
-				: modeItems[0]?.value;
-
-			modeItems.forEach((item, index) => {
-				sheetItems.push({
-					id: registerAction({ kind: 'agentHostMode', value: item.value }),
-					label: item.label,
-					description: item.description,
-					icon: getAgentHostModeIcon(item.value),
-					checked: item.value === currentModeValue,
-					sectionTitle: index === 0
-						? localize('chatPhoneInput.modeSection', "Agent Mode")
-						: undefined,
-				});
-			});
-
-			// Filter the language models by the session's resource
-			// scheme — same logic as `getAgentHostModels` in
-			// `agentHostModelPicker.ts` and `mobileChatInputConfigPicker.ts`.
-			const resourceScheme = activeSession.resource.scheme;
-			const agentHostModels = this._languageModelsService.getLanguageModelIds()
-				.map(id => {
-					const metadata = this._languageModelsService.lookupLanguageModel(id);
-					return metadata ? { metadata, identifier: id } : undefined;
-				})
-				.filter((m): m is ILanguageModelChatMetadataAndIdentifier => !!m && m.metadata.targetChatSessionType === resourceScheme);
-			// Match desktop `agentHostModelPicker.ts`: only fall back to
-			// the per-scheme storage key for untitled sessions. A saved
-			// session has its own `modelId`.
-			const isUntitled = activeSession.status.get() === SessionStatus.Untitled;
-			const storedModelId = isUntitled
-				? this._storageService.get(agentHostModelPickerStorageKey(resourceScheme), StorageScope.PROFILE)
-				: undefined;
-			const currentModelId = activeSession.modelId.get() ?? storedModelId;
-
-			agentHostModels.forEach((model, index) => {
-				sheetItems.push({
-					id: registerAction({ kind: 'agentHostModel', model }),
-					label: model.metadata.name,
-					icon: getModelProviderIcon(model),
-					checked: model.identifier === currentModelId,
-					sectionTitle: index === 0
-						? localize('chatPhoneInput.modelSection', "Model")
-						: undefined,
-				});
-			});
+			sheetItems = this._buildAgentHostSheetItems(activeSession, agentHostProvider, registerAction);
 		} else {
-			// Default Copilot chat path: requires the workbench
-			// delegates. Callers without delegates (e.g. the agent-host
-			// mode pill on phone, when no agent-host session is active)
-			// have nothing to show.
 			if (!modeDelegate || !modelDelegate) {
 				return;
 			}
-			const modes = modeDelegate.currentChatModes.get();
-			const currentMode = modeDelegate.currentMode.get();
-			const modelItems = modelDelegate.getModels();
-			const currentModel = modelDelegate.currentModel.get();
-
-			const allModes = [...modes.builtin, ...modes.custom];
-			allModes.forEach((mode, index) => {
-				const icon = mode.icon.get();
-				sheetItems.push({
-					id: registerAction({ kind: 'mode', mode }),
-					label: mode.label.get(),
-					icon: ThemeIcon.isThemeIcon(icon) ? icon : undefined,
-					checked: mode.id === currentMode.id,
-					sectionTitle: index === 0
-						? localize('chatPhoneInput.modeSection', "Agent Mode")
-						: undefined,
-				});
-			});
-
-			modelItems.forEach((model, index) => {
-				sheetItems.push({
-					id: registerAction({ kind: 'model', model }),
-					label: model.metadata.name,
-					icon: getModelProviderIcon(model),
-					checked: model.identifier === currentModel?.identifier,
-					sectionTitle: index === 0
-						? localize('chatPhoneInput.modelSection', "Model")
-						: undefined,
-				});
-			});
+			sheetItems = this._buildDelegateSheetItems(modeDelegate, modelDelegate, registerAction);
 		}
 
 		if (sheetItems.length === 0) {
 			return;
 		}
-
-		const performAction = (action: ChatPhonePickerAction): void => {
-			// Re-resolve the active session and its provider on every
-			// tap. The sheet stays open across multiple selections via
-			// `stayOpenOnSelect`, and the active session can change
-			// while it's open (e.g. background switch). Capturing once
-			// at sheet-open would silently apply later writes to the
-			// stale session.
-			const session = this._sessionsService.activeSession.get();
-			const provider = session ? this._sessionsProvidersService.getProvider(session.providerId) : undefined;
-			const ahProvider = provider && isAgentHostProvider(provider) ? provider : undefined;
-
-			switch (action.kind) {
-				case 'mode':
-					// Same dispatch the desktop mode picker uses (see
-					// `modePickerActionItem.ts` — the row's `run()` invokes
-					// `ToggleAgentModeActionId` with `{ modeId, sessionResource }`).
-					this._commandService.executeCommand(
-						ToggleAgentModeActionId,
-						{ modeId: action.mode.id, sessionResource: modeDelegate?.sessionResource() } satisfies IToggleChatModeArgs,
-					).catch(() => { /* best-effort */ });
-					break;
-				case 'model':
-					// Same dispatch the desktop model picker uses (see
-					// `ModelPickerActionItem` — `onDidChangeSelection` routes to
-					// `delegate.setModel`).
-					modelDelegate?.setModel(action.model);
-					break;
-				case 'agentHostMode':
-					// Same write path as `MobileChatInputConfigPicker` and
-					// `AgentHostModePicker._showPicker`'s `onSelect`.
-					if (session && ahProvider) {
-						ahProvider.setSessionConfigValue(session.sessionId, SessionConfigKey.Mode, action.value)
-							.catch(() => { /* best-effort */ });
-					}
-					break;
-				case 'agentHostModel':
-					// Drive the workbench delegate (when present) so the
-					// chip's `currentModel` observable updates immediately
-					// and the input toolbar repaints.
-					if (modelDelegate) {
-						modelDelegate.setModel(action.model);
-					} else if (session) {
-						// Caller (e.g. the agent-host mode pill) didn't
-						// pass a delegate. Look up the chat widget by
-						// the active session's resource — this is the
-						// input that owns the `_currentLanguageModel`
-						// observable the workbench chip reads. Using
-						// `lastFocusedWidget` would be a global focus
-						// tracker and could push the chip update into a
-						// different input than the model write.
-						this._chatWidgetService.getWidgetBySessionResource(session.resource)
-							?.input.setCurrentLanguageModel(action.model, true);
-					}
-					if (session && ahProvider) {
-						// Persist to the shared storage key so the empty
-						// new-chat picker (`MobileChatInputConfigPicker`)
-						// remembers the same selection across surfaces,
-						// and push to the agent-host provider so the
-						// next send goes out with the picked model.
-						setAgentHostModelSelection(session, [action.model], action.model.identifier, ahProvider, this._storageService);
-					}
-					break;
-			}
-		};
 
 		// Use `stayOpenOnSelect` so tapping a row applies the choice but
 		// keeps the sheet visible until the user explicitly hits Done
@@ -301,11 +128,131 @@ class MobileChatPhoneInputPresenter extends Disposable implements IChatPhonePres
 				onDidSelect: id => {
 					const action = idToAction.get(id);
 					if (action) {
-						performAction(action);
+						this._performAction(action, target, modeDelegate, modelDelegate);
 					}
 				},
 			},
 		);
+	}
+
+	private _buildAgentHostSheetItems(
+		session: IActiveSession,
+		provider: IAgentHostSessionsProvider,
+		registerAction: RegisterChatPhonePickerAction,
+	): IMobilePickerSheetItem[] {
+		const items: IMobilePickerSheetItem[] = [];
+		const config = provider.getSessionConfig(session.sessionId);
+		const modeSchema = config?.schema.properties[SessionConfigKey.Mode];
+		const modeItems = (modeSchema && isWellKnownModeSchema(modeSchema))
+			? (modeSchema.enum ?? []).map((value, index) => ({
+				value: String(value),
+				label: modeSchema.enumLabels?.[index] ?? String(value),
+				description: modeSchema.enumDescriptions?.[index],
+			}))
+			: [];
+		const rawCurrentMode = config?.values[SessionConfigKey.Mode] ?? modeSchema?.default;
+		const currentModeValue = typeof rawCurrentMode === 'string' && modeItems.some(item => item.value === rawCurrentMode)
+			? rawCurrentMode
+			: modeItems[0]?.value;
+
+		modeItems.forEach((item, index) => items.push({
+			id: registerAction({ kind: 'agentHostMode', value: item.value }),
+			label: item.label,
+			description: item.description,
+			icon: getAgentHostModeIcon(item.value),
+			checked: item.value === currentModeValue,
+			sectionTitle: index === 0 ? localize('chatPhoneInput.modeSection', "Agent Mode") : undefined,
+		}));
+
+		const models = provider.getModels(session.sessionId);
+		const currentModelId = session.modelId.get();
+		models.forEach((model, index) => items.push({
+			id: registerAction({ kind: 'agentHostModel', model }),
+			label: model.metadata.name,
+			icon: getModelProviderIcon(model),
+			checked: model.identifier === currentModelId,
+			sectionTitle: index === 0 ? localize('chatPhoneInput.modelSection', "Model") : undefined,
+		}));
+
+		const options = normalizeModelPickerOptions(provider.getModelPickerOptions(session.sessionId));
+		if (models.length === 0 && !options.showAutoModel) {
+			items.push({
+				id: 'chat-phone-picker-no-models',
+				label: localize('chatPhoneInput.noModels', "No models available"),
+				disabled: true,
+				sectionTitle: localize('chatPhoneInput.modelSection', "Model"),
+			});
+		}
+		return items;
+	}
+
+	private _buildDelegateSheetItems(
+		modeDelegate: IModePickerDelegate,
+		modelDelegate: IModelPickerDelegate,
+		registerAction: RegisterChatPhonePickerAction,
+	): IMobilePickerSheetItem[] {
+		const items: IMobilePickerSheetItem[] = [];
+		const modes = modeDelegate.currentChatModes.get();
+		const currentMode = modeDelegate.currentMode.get();
+		[...modes.builtin, ...modes.custom].forEach((mode, index) => {
+			const icon = mode.icon.get();
+			items.push({
+				id: registerAction({ kind: 'mode', mode }),
+				label: mode.label.get(),
+				icon: ThemeIcon.isThemeIcon(icon) ? icon : undefined,
+				checked: mode.id === currentMode.id,
+				sectionTitle: index === 0 ? localize('chatPhoneInput.modeSection', "Agent Mode") : undefined,
+			});
+		});
+
+		const currentModel = modelDelegate.currentModel.get();
+		modelDelegate.getModels().forEach((model, index) => items.push({
+			id: registerAction({ kind: 'model', model }),
+			label: model.metadata.name,
+			icon: getModelProviderIcon(model),
+			checked: model.identifier === currentModel?.identifier,
+			sectionTitle: index === 0 ? localize('chatPhoneInput.modelSection', "Model") : undefined,
+		}));
+		return items;
+	}
+
+	private _performAction(
+		action: ChatPhonePickerAction,
+		target: IChatPhoneInputTarget | undefined,
+		modeDelegate: IModePickerDelegate | undefined,
+		modelDelegate: IModelPickerDelegate | undefined,
+	): void {
+		const session = this._sessionsService.activeSession.get();
+		if (!matchesChatPhoneInputTarget(target, session)) {
+			return;
+		}
+		const provider = session ? this._sessionsProvidersService.getProvider(session.providerId) : undefined;
+		const agentHostProvider = provider && isAgentHostProvider(provider) ? provider : undefined;
+
+		switch (action.kind) {
+			case 'mode':
+				this._commandService.executeCommand(
+					ToggleAgentModeActionId,
+					{ modeId: action.mode.id, sessionResource: modeDelegate?.sessionResource() } satisfies IToggleChatModeArgs,
+				).catch(() => { });
+				break;
+			case 'model':
+				modelDelegate?.setModel(action.model);
+				break;
+			case 'agentHostMode':
+				if (session && agentHostProvider) {
+					const schema = agentHostProvider.getSessionConfig(session.sessionId)?.schema.properties[SessionConfigKey.Mode];
+					if (schema && isWellKnownModeValue(schema, action.value)) {
+						agentHostProvider.setSessionConfigValue(session.sessionId, SessionConfigKey.Mode, action.value).catch(() => { });
+					}
+				}
+				break;
+			case 'agentHostModel':
+				if (session && agentHostProvider) {
+					selectAvailableSessionModel(session, agentHostProvider, this._storageService, action.model.identifier);
+				}
+				break;
+		}
 	}
 }
